--- src/components/BackButton/index.tsx
@@ -1,1 +1,0 @@
-export { BackButton } from './BackButton';
