--- src/components/Main/index.tsx
@@ -1,1 +1,0 @@
-export { Main } from './Main';
