--- src/components/Navigation/index.tsx
@@ -1,1 +1,0 @@
-export { Navigation } from './Navigation';
