--- src/components/ProductItem/index.tsx
@@ -1,1 +1,0 @@
-export { ProductItem } from './ProductItem';
