--- src/components/Products/index.tsx
@@ -1,1 +1,0 @@
-export { Products } from './Products';
