--- src/components/RatingLogo/index.tsx
@@ -1,1 +1,0 @@
-export { RatingLogo } from './RatingLogo';
