--- src/components/SearchForm/index.tsx
@@ -1,1 +1,0 @@
-export { SearchForm } from './SearchForm';
