+++ src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { Products } from '../components/Products';
+import { Products } from '../components/Products/Products';
 
 class Home extends Component {
   render() {
