+++ src/pages/Layout.tsx
@@ -1,8 +1,8 @@
 import { Component } from 'react';
 import { Outlet } from 'react-router-dom';
-import { Header } from '../components/Header';
-import { Footer } from '../components/Footer';
-import { Main } from '../components/Main';
+import { Header } from '../components/Header/Header';
+import { Footer } from '../components/Footer/Footer';
+import { Main } from '../components/Main/Main';
 
 class Layout extends Component {
   render() {
